requirements.txt
fastapi
psycopg2===2.9.9
psycopg2-binary===2.9.9
python-dotenv===1.0.1
uvicorn===0.30.1